--- frontend/src/components/Header.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-import { Menu } from "semantic-ui-react";
-
-export const Header = () => (
-  <Menu fluid color="grey">
-    <Menu.Item header>Key Conjurer</Menu.Item>
-    <Menu.Item position="right">{process.env.REACT_APP_VERSION}</Menu.Item>
-  </Menu>
-);
